Article warning fixes
