+++ Tests/LineFactoryTest.php
@@ -5,7 +5,6 @@ namespace NumaxLab\Geslib\Tests;
 use NumaxLab\Geslib\Exceptions\InvalidLineCodeException;
 use NumaxLab\Geslib\Exceptions\NotImplementedLineCodeException;
 use NumaxLab\Geslib\LineFactory;
-use NumaxLab\Geslib\Lines\ArtAdv;
 use NumaxLab\Geslib\Lines\BindingType;
 use PHPUnit\Framework\TestCase;
 
@@ -34,18 +33,5 @@ class LineFactoryTest extends TestCase
         $this->expectException(NotImplementedLineCodeException::class);
 
         LineFactory::create('PACK', []);
-    }
-
-    public function testCreatesArtAdvLine()
-    {
-        $artAdvLine = LineFactory::create(ArtAdv::CODE, [
-            ArtAdv::CODE,
-            'A',
-            '123',
-            '456',
-            'Test Description',
-        ]);
-
-        $this->assertInstanceOf(ArtAdv::class, $artAdvLine);
     }
 }
